requirements.txt
hydrogram>=0.2.0
peewee==3.16.0
pysocks==1.7.1
python-dotenv==1.0.0
python-environ>=0.4.54
python-telegram-bot[ext]>=22.5
uvloop>=0.21.0
